Fixed build. Deactivated CpuCyclesSpentSelf to reduce the number of queries.
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@5017 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/AddIns/Misc/Profiler/Controller/Data/ProfilingDataSQLiteProvider.cs
@@ -311,7 +311,7 @@ namespace ICSharpCode.Profiler.Controller.Data
 					while (reader.Read()) {
 						SQLiteCallTreeNode node = new SQLiteCallTreeNode(reader.GetInt32(1), null, queryProvider);
 						node.callCount = reader.GetInt32(3);
-						node.cpuCyclesSpent = (ulong)reader.GetInt64(2);
+						node.cpuCyclesSpent = reader.GetInt64(2);
 						object ids = reader.GetValue(0);
 						if (ids is long) {
 							node.ids = new List<int> { (int)(long)ids };
